requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
requests>=2.31.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
